net: fix concurrency bugs, crash-safety, and dedup across the net/ layer
Code-review sweep: every finding implemented.

Concurrency:
- RingBuffer: one FrameRingBuffer<N> template (replaces StereoRingBuffer +
  NjAudio's WideRing copy) with consumer-applied requestClear()/drain() —
  resetting indices from the UI thread raced a concurrent pull() and could
  replay up to a full ring of the previous stream. TX capture rings are now
  drained by their consumer (txLoop), also giving clean downbeat-aligned
  intervals on TX re-enable.
- StreamClient/NjClient: sockMutex makes stop()'s shutdown() atomic with the
  bg thread's close() (no shutdown of a recycled fd).
- netConnectAbortable() (Socket.cpp): shared 100ms-abort-poll connect used by
  Http, StreamClient, and NjClient — NjClient's blocking connect() could hang
  stop()/join on the UI thread for the full OS timeout.
- NjClient TX: txChannels behind txMutex; onUploadInterval bound in start()
  only; Ninjam's isOwnSpeaker uses a per-session captured name; credentials
  assigned only after stopAll() (joinManual takes user/pass).
- Ninjam: cross-thread flags (mode/listening/joined/clickEnabled/transmitting/
  txQuality/txDeclared) are atomics; process() reads them relaxed once/sample.

Crash-safety:
- jsonStr() guards every json_string_value read (malformed patch => no crash).
- SET_USERMASK: guard channelIdx >= 32 (UB shift on hostile wire data).
- ImageCache ICO: 64-bit bounds check (uint32 off+size overflow -> substr throw).
- RoomDirectory/StationImporter destructors abort their fetch before joining
  (module delete no longer stalls the UI for the HTTP timeout).
- MP3: PCM buffer always stereo-sized; channels/sampleRate re-read per block
  (mid-stream format change can't overflow; pitch stays correct).
- NjAudio: cap intervalSamples (hostile tempo can't size decode buffers to GBs).

Behavior/efficiency:
- Stream socket gets SO_RCVTIMEO + a 30s idle budget (readIdle): a stalled-but-
  open stream now ends with "Stream ended" instead of hanging in recv forever.
- One ResampleCtx replaces three duplicated resampler+backpressure copies;
  shared parseUrl/endsWithCI (Http.hpp); shared flexInt/flexStr (net/Json.hpp);
  StationImport uses the real StreamClient::State enum; shared SSL_CTX;
  httpGet fails on >4MiB instead of returning a silent truncation; pow->v*v;
  resyncBeat relaxed-load fast path; dead PLAYING_LIGHT removed.

Docs: stale "no transmit yet" comments rewritten (NjClient.hpp, Ninjam.cpp);
ring-headroom comment corrected (1.4 s); play_test link line gains Socket.cpp.

Verified: clean rebuild (only pre-existing vendored warnings); make leakcheck
=> enc_test + leak_stress (live MP3 + AAC/HLS churn x3) report 0 leaks.

+++ src/net/ImageCache.cpp
@@ -87,7 +87,9 @@ bool icoToImage(const std::string& b, std::string& out, std::string& ext) {
 		uint16_t bpp = rd16(b, e + 6);
 		uint32_t size = rd32(b, e + 8);
 		uint32_t off = rd32(b, e + 12);
-		if (size == 0 || off + size > b.size())
+		// Bounds-check in 64-bit: off + size can wrap uint32 on hostile input,
+		// which would pass the check and make substr() below throw.
+		if (size == 0 || (uint64_t) off + (uint64_t) size > (uint64_t) b.size())
 			continue;
 		long score = (long) w * h * 100 + bpp;
 		if (score > bestScore) {
